修正baidu的seek问题；对角色和速度的本地记录

+++ app/src/main/java/cn/xylink/mting/MTing.java
@@ -29,12 +29,14 @@ import cn.xylink.mting.model.data.OkGoUtils;
 import cn.xylink.mting.openapi.QQApi;
 import cn.xylink.mting.openapi.WXapi;
 import cn.xylink.mting.speech.SpeechService;
+import cn.xylink.mting.speech.Speechor;
 import cn.xylink.mting.upgrade.UpgradeManager;
 import cn.xylink.mting.utils.ContentManager;
 import cn.xylink.mting.utils.EncryptionUtil;
 import cn.xylink.mting.utils.GsonUtil;
 import cn.xylink.mting.utils.ImageUtils;
 import cn.xylink.mting.utils.PackageUtils;
+import cn.xylink.mting.utils.SharedPreHelper;
 import okhttp3.OkHttpClient;
 
 public class MTing extends Application {
@@ -70,7 +72,12 @@ public class MTing extends Application {
         ImageUtils.init(this);
 
         clearAudioCache();
-        startService(new Intent(this, SpeechService.class));
+        Intent serviceIntent = new Intent(this, SpeechService.class);
+        String defaultRole = String.valueOf(SharedPreHelper.getInstance(this).getSharedPreference("SPEECH_ROLE", "XiaoIce"));
+        String defaultSpeed = String.valueOf(SharedPreHelper.getInstance(this).getSharedPreference("SPEECH_SPEED", "SPEECH_SPEED_NORMAL"));
+        serviceIntent.putExtra("role", defaultRole);
+        serviceIntent.putExtra("speed", defaultSpeed);
+        startService(serviceIntent);
 
         try {
             checkOnlineUpgrade();
@@ -107,6 +114,9 @@ public class MTing extends Application {
         AudioCachePath = audioCacheFile.getPath();
         new Thread(()->{
             File[] mp3Files = audioCacheFile.listFiles();
+            if(mp3Files == null || mp3Files.length <= 0) {
+                return;
+            }
             for(File mp3File : mp3Files) {
                 mp3File.delete();
             }
